OFC-backend/ml/requirements.txt
# forticrypt-backend/ml/requirements.txt
# Core ML dependencies
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3

# File analysis dependencies
python-magic>=0.4.27
python-magic-bin>=0.4.14  # For Windows compatibility
